# LCDIF and DCNANO performance improvements for LVGL