--- a/boards/arm/mimxrt1060_evk/Kconfig.defconfig
+++ b/boards/arm/mimxrt1060_evk/Kconfig.defconfig
@@ -1,6 +1,6 @@
 # MIMXRT1060-EVK board
 
-# Copyright (c) 2018, NXP
+# Copyright 2018,2023 NXP
 # SPDX-License-Identifier: Apache-2.0
 
 if BOARD_MIMXRT1060_EVK || BOARD_MIMXRT1060_EVK_HYPERFLASH || BOARD_MIMXRT1060_EVKB
@@ -56,11 +56,29 @@ if LVGL
 config LV_Z_POINTER_KSCAN
 	default y
 
+# LVGL should allocate buffers equal to size of display
 config LV_Z_VDB_SIZE
-	default 16
+	default 100
+
+# Enable double buffering
+config LV_Z_DOUBLE_VDB
+	default y
+
+# Force full refresh. This prevents memory copy associated with partial
+# display refreshes, which is not necessary for the eLCDIF driver
+config LV_Z_FULL_REFRESH
+	default y
 
 config LV_Z_DPI
 	default 128
+
+config LV_Z_BITS_PER_PIXEL
+	default 16
+
+# Force display buffers to be aligned to cache line size (32 bytes)
+config LV_Z_VDB_ALIGN
+	default 32
+
 
 choice LV_COLOR_DEPTH
 	default LV_COLOR_DEPTH_16