--- a/boards/arm/mimxrt595_evk/Kconfig.defconfig
+++ b/boards/arm/mimxrt595_evk/Kconfig.defconfig
@@ -41,6 +41,8 @@ config MCUX_DCNANO_LCDIF_EXTERNAL_FB_MEM
 	default y
 # Use FlexSPI2 base address for framebuffer (pSRAM is present on this bus)
 config MCUX_DCNANO_LCDIF_EXTERNAL_FB_ADDR
+	# Move DCNANO framebuffer if LVGL framebuffers are also in PSRAM
+	default 0x38400000 if LV_Z_VBD_CUSTOM_SECTION
 	default 0x38000000
 # M33 core and LCDIF both access FlexSPI2 through the same cache,
 # so coherency does not need to be managed.
@@ -53,6 +55,34 @@ config KSCAN
 	default y if LVGL
 
 if LVGL
+
+# LVGL should allocate buffers equal to size of display
+config LV_Z_VDB_SIZE
+	default 100
+
+# Enable double buffering
+config LV_Z_DOUBLE_VDB
+	default y
+
+# Force full refresh. This prevents memory copy associated with partial
+# display refreshes, which is not necessary for the eLCDIF driver
+config LV_Z_FULL_REFRESH
+	default y
+
+config LV_Z_DPI
+	default 128
+
+config LV_Z_BITS_PER_PIXEL
+	default 16
+
+# Force display buffers to be aligned for DCNANO LCDIF (128 byte alignment)
+config LV_Z_VDB_ALIGN
+	default 128
+
+# LVGL display buffers will be too large for internal SRAM, locate in
+# custom section within PSRAM
+config LV_Z_VBD_CUSTOM_SECTION
+	default y
 
 config LV_Z_POINTER_KSCAN
 	default y