This PR includes the following updates:

- Updates Zephyr's revision of LVGL to include performance tuning options for the following settings:
    - Enable forcing LVGL to perform a full display refresh on every update
    - Enable custom alignment and linker section location of LVGL framebuffers.
- Updates MCUX ECLDIF and DCNANO drivers to pass the framebuffer in the `display_write` call directly to hardware when possible (ie when the buffer is the same size as the display). This enables the calling application to avoid the expensive memory copies associated with partial display updates by always calling `display_write` with a buffer the same size as the connected display
- Updates LVGL configuration for the RT595, RT1060, and RT1170 EVK with the following settings:
    - Allocate dual LVGL rendering buffers equal in size to the connected display
    - Always refresh the entire display (so that the hardware can use the LVGL rendering buffer directly)
    - For the RT595 EVK, relocate the LVGL rendering buffers to external PSRAM since they will not fit in internal RAM


This results in the following performance improvements in the weighted FPS of the LVGL benchmark sample:

RT595:
3 FPS -> 7 FPS

RT1170:
3 FPS -> 34 FPS

RT1060:
18 FPS -> 65 FPS